+++ src/entrypoint.rb
@@ -1,4 +1,5 @@
 
 $: << 'src'
+$: << 'src/libs/aniruby'
 
 require 'main'