+++ src/libs/aniruby/aniruby.rb
@@ -1,0 +1,6 @@
+
+require_relative 'aniruby/version'
+
+require_relative 'aniruby/frame'
+require_relative 'aniruby/frames'
+require_relative 'aniruby/animation'